requirements.txt
# Core dependencies
streamlit>=1.28.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# API & Data Processing
requests>=2.31.0
xmltodict>=0.13.0
pydantic>=2.0.0
aiohttp>=3.9.0

# Security
python-jose>=3.3.0  # JWT tokens
bcrypt>=4.0.0       # Password hashing
passlib>=1.7.4

# Performance
redis>=5.0.0        # Optional - for caching
asyncpg>=0.29.0     # Async PostgreSQL driver

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Monitoring
loguru>=0.7.0       # Better logging